feat: Mejora de la gestión de paquetes y actividades con la opción de asociar destinos y formatear precios.
- Actualiza `PaqueteService` para permitir la asociación opcional de destinos al crear un paquete.
- Modifica `ActividadDTO`, `DestinoDTO`, `PagoDTO`, `PaqueteDTO` y `ReservaDTO` para usar números enteros en los precios, reflejando así la moneda chilena.
- Mejora la interfaz de usuario en los menús de administración para destinos y actividades, incluyendo un mejor manejo de entradas y visualización de las opciones disponibles.
- Añade funciones de utilidad para depurar entradas numéricas y calcular precios con IVA.
- Mejora las funciones de visualización de tablas para incluir el IVA en los precios cuando corresponda.
- Corrige varios errores tipográficos y se mejora la legibilidad del código en varios archivos.

+++ src/dto/actividad_dto.py
@@ -1,6 +1,7 @@
 class ActividadDTO:
-    # Clase que representa los datos de una actividad turística.
-    def __init__(self, id:int | None, nombre:str, descripcion: str, duracion_horas: int, precio_base: float, destino_id:int):
+    # Clase que representa los datos de una actividad turistica.
+    # Precios en pesos chilenos (enteros, sin decimales)
+    def __init__(self, id:int | None, nombre:str, descripcion: str, duracion_horas: int, precio_base: int, destino_id:int):
         self.id=id
         self.nombre=nombre
         self.descripcion=descripcion
