+++ src/dto/pago_dto.py
@@ -3,7 +3,8 @@ from datetime import datetime
 
 class PagoDTO:
     # Clase que representa los datos de un pago.
-    def __init__(self, id:int, monto:float, fecha_pago:datetime, metodo:str, reserva_id:int, estado:str):
+    # Montos en pesos chilenos (enteros, sin decimales)
+    def __init__(self, id:int, monto:int, fecha_pago:datetime, metodo:str, reserva_id:int, estado:str):
         self.id=id
         self.monto=monto
         self.fecha_pago=fecha_pago
